requirements.txt
slack-sdk==3.24.0
openai==1.3.7
python-dotenv==1.0.0
pydantic==2.5.0
rich==13.7.0
typer==0.9.0
